Intrroduce explict flags for testing mode
This revision introduces reSeed and predResistance flags to define the
mode of testing and the corresponding data. See also the table in
Section 3.1 The array otherInput contains the test data for the
configured mode of testing. Added a new example for the case of reSeed
and predResistance both set to "no".