This revision commits an initial framework for algorithm test extensions. The minimum requires set to explain this consists of three specs: the protocol, the new algorithm test extension spec, the symmetric block cipher spec.
This is only the first iteration on this and as such it is incomplete in
elaborating fully on all the necessary text. However, it allows the effort
to be partitioned between the editors for the three documents.

Note also that the new specs are renamed to match IETF guidelines for
Internet drafts.