This commit creates an initial IANA registry for ACVP algorithm test extensions. The proposed registry is hierarchical, starting with algorithm classes that contain algorithm types that contain algorithms.
The algorithm class registry is fully defined. The algorithm types of only a few classes are defined. Only the algorithms for symmetric block ciphers are defined at this stage. This is on purpose to keep the registry fully defined for the envisioned first step of covering with the specification only the symmetric block ciphers.