Fix swapped naming of test case fields
In the present version of the spec, table 5 names the fields of the ACVP
test cases. However, the order of the parts of the name (namely,
hello/finished and server/client) are swapped from the value given in
the example and by the Demo server. E.g., serverHelloRandom is in
practice called helloServerRandom. Fix the two to match.

Signed-off-by: Alexander Scheel <alex.scheel@hashicorp.com>